Some Issues Fixed

+++ IkostProjeMVC/Models/Entities/Bayiler.cs
@@ -6,7 +6,7 @@ namespace IkostProjeMVC.Models.Entities
     {
         public int BayiID { get; set; }
         public string BayiAd { get; set; }
-        public int HakedisYuzde { get; set; }
+        public decimal HakedisYuzde { get; set; }
         public List<Siparisler> Siparislers { get; set; }
     }
 }
